requirements.txt
asgiref==3.5.1
async-timeout==4.0.2
charset-normalizer==2.0.12
Deprecated==1.2.13
Django==3.2.13
django-cors-headers==3.7.0
django-filter==21.1
django-storages==1.13.2
djangorestframework==3.12.4
idna==3.3
jmespath==1.0.1
packaging==23.0
Pillow==9.1.0
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
requests==2.27.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
uWSGI==2.0.20
wrapt==1.14.1
xmltodict==0.13.0